Rover: cleanup build warnings

+++ APMrover2/sensors.pde
@@ -60,8 +60,8 @@ static void read_sonars(void)
         // we have two sonars
         obstacle.sonar1_distance_cm = sonar.distance_cm();
         obstacle.sonar2_distance_cm = sonar2.distance_cm();
-        if (obstacle.sonar1_distance_cm <= g.sonar_trigger_cm &&
-            obstacle.sonar2_distance_cm <= obstacle.sonar2_distance_cm)  {
+        if (obstacle.sonar1_distance_cm <= (uint16_t)g.sonar_trigger_cm &&
+            obstacle.sonar2_distance_cm <= (uint16_t)obstacle.sonar2_distance_cm)  {
             // we have an object on the left
             if (obstacle.detected_count < 127) {
                 obstacle.detected_count++;
@@ -72,7 +72,7 @@ static void read_sonars(void)
             }
             obstacle.detected_time_ms = hal.scheduler->millis();
             obstacle.turn_angle = g.sonar_turn_angle;
-        } else if (obstacle.sonar2_distance_cm <= g.sonar_trigger_cm) {
+        } else if (obstacle.sonar2_distance_cm <= (uint16_t)g.sonar_trigger_cm) {
             // we have an object on the right
             if (obstacle.detected_count < 127) {
                 obstacle.detected_count++;
@@ -88,7 +88,7 @@ static void read_sonars(void)
         // we have a single sonar
         obstacle.sonar1_distance_cm = sonar.distance_cm();
         obstacle.sonar2_distance_cm = 0;
-        if (obstacle.sonar1_distance_cm <= g.sonar_trigger_cm)  {
+        if (obstacle.sonar1_distance_cm <= (uint16_t)g.sonar_trigger_cm)  {
             // obstacle detected in front 
             if (obstacle.detected_count < 127) {
                 obstacle.detected_count++;
